Update RB on 211105.md
First little remarks  ;)  Problem I have with some comments are that some things (in public sector) are self-evident if you are in that sector. 

Example:
Yes governments are (generic) split into three layers. As far as I know people have not done research on it. Its just the way it is ;) It is not on me to do this research. But the distributions and the laws can differ per country. But this is also self-evident (not researched as far as I know). Example we (NL) had that strange split in our police (regional, Central and local) that you do not see everywhere.  Regional is gone btw.. that is removed 20 years ago so we have now central and local. But in other countries they do it differently. But still there is a Central, Regional and Local distinction. The distribution is regulated by law and policies created by parliament and can change every moment (stressor on our public sector).

Summary some things are just the way they are ;) (and I and Hans agreed that I would stay away as far as possible on the Public Administration research. It is only a context.