First concept of expert group chapter
The first concept of the expert group chapter is finished.

Also added Antwerp, Belgium to the title page.

Changed the footnote definition for a better readability
Moved data set of validation to the main repository for referencing
Corrected the ToC added to the ToC (page number was +1 off)
Removed appendices that are not necessary anymore. Corrected some appendices to make them better readable.
Changed the bib file based on the feedback received. Solved some issues with the references.
Added ISBN to the references (non-APA Standard). Added links to all the books in the reference (Amazon and when not found on an international site WorldCat is used). URLs are optional with APA.